(#511) Tests: extract PreprocessorUtil
