+++ Cesium.TestFramework/PreprocessorUtil.cs
@@ -1,34 +1,20 @@
+using System.Diagnostics.CodeAnalysis;
 using Cesium.Core.Warnings;
 using Cesium.Preprocessor;
-using Cesium.TestFramework;
-using JetBrains.Annotations;
-using System.Diagnostics.CodeAnalysis;
 using Yoakke.SynKit.Lexer;
 
-namespace Cesium.CodeGen.Tests;
+namespace Cesium.TestFramework;
 
-public class CodeGenPinvokeTests : CodeGenTestBase
+public static class PreprocessorUtil
 {
-    private const string _mainMockedFilePath = @"c:\a\b\c.c";
-
-    [MustUseReturnValue]
-    private static Task DoTest(string source)
-    {
-        var processed = DoPreprocess(source, null, null, null);
-        processed.Wait();
-        var assembly = GenerateAssembly(default, processed.Result);
-
-        var moduleType = assembly.Modules.Single().GetType("<Module>");
-        return VerifyMethods(moduleType);
-    }
-
-    private static async Task<string> DoPreprocess(
+    public static async Task<string> DoPreprocess(
+        string sourceFileName,
         [StringSyntax("cpp")] string source,
         Dictionary<string, string>? standardHeaders = null,
         Dictionary<string, IList<IToken<CPreprocessorTokenType>>>? defines = null,
         Action<PreprocessorWarning>? onWarning = null)
     {
-        var lexer = new CPreprocessorLexer(_mainMockedFilePath, source);
+        var lexer = new CPreprocessorLexer(sourceFileName, source);
         var includeContext = new IncludeContextMock(standardHeaders ?? new Dictionary<string, string>());
         var definesContext = new InMemoryDefinesContext();
         if (defines != null)
@@ -45,7 +31,7 @@ public class CodeGenPinvokeTests : CodeGenTestBase
         using (warningProcessor as IDisposable)
         {
             var preprocessor = new CPreprocessor(
-                _mainMockedFilePath,
+                sourceFileName,
                 lexer,
                 includeContext,
                 definesContext,
@@ -54,23 +40,4 @@ public class CodeGenPinvokeTests : CodeGenTestBase
             return result;
         }
     }
-
-    [Fact]
-    public Task SinglePinvokePragma() => DoTest(@"
-#pragma pinvoke(""mydll.dll"")
-int not_pinvoke(void);
-int foo_bar(int*);
-
-int main() {
-    return foo_bar(0);
-}
-
-int not_pinvoke(void) { return 1; }
-");
-
-    [Fact] // win_puts -> pinvokeimpl(msvcrt, puts) int win_puts();
-    public Task PinvokePrefixPragma() => DoTest(@"
-#pragma pinvoke(""msvcrt"", win_)
-int win_puts(const char*);
-");
 }
